pyproject.toml
[tool.poetry]
name = "barbershop-django"
version = "0.1.0"
description = "Барбершоп 'Стальная Борода' - система онлайн-записи на Django"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "barbershop"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "barbershop.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]